testing to compile usImageRf2D with vpImage<double> instead of <short>

+++ modules/ustk_core/include/visp3/ustk_core/usImageRF2D.h
@@ -49,7 +49,7 @@
  *
  * This class represents a 2D ultrasound RF frame.
  */
-class VISP_EXPORT usImageRF2D : public vpImage<short>, public usImageSettings {
+class VISP_EXPORT usImageRF2D : public vpImage<double>, public usImageSettings {
  public:
   /**unsigned char
    * Constructor.
