+++ modules/ustk_core/src/usImageRF2D.cpp
@@ -1,13 +1,13 @@
 #include <visp3/ustk_core/usImageRF2D.h>
 #include <cstring>
 
-usImageRF2D::usImageRF2D() : vpImage<short>(), usImageSettings()
+usImageRF2D::usImageRF2D() : vpImage<double>(), usImageSettings()
 {
 	
 }
 
 usImageRF2D::usImageRF2D(unsigned int AN, unsigned int LN)
-  : vpImage<short>(AN, LN), usImageSettings()
+  : vpImage<double>(AN, LN), usImageSettings()
 {
 	
 }
@@ -15,13 +15,13 @@ usImageRF2D::usImageRF2D(unsigned int AN, unsigned int LN)
 usImageRF2D::usImageRF2D(unsigned int AN, unsigned int LN,
 		       float probeRadius, float lineAngle, float resolution,
 		       float BSampleFreq, float probeElementPitch)
-  : vpImage<short>(AN, LN), usImageSettings(probeRadius, lineAngle, resolution, BSampleFreq, probeElementPitch)
+  : vpImage<double>(AN, LN), usImageSettings(probeRadius, lineAngle, resolution, BSampleFreq, probeElementPitch)
 {
 	
 }
 
 usImageRF2D::usImageRF2D(const usImageRF2D& other)
-  : vpImage<short>(other), usImageSettings()
+  : vpImage<double>(other), usImageSettings()
 {
 	
 }
